requirements.txt
flask
python-oracledb
